requirements.txt
Flask==3.0.0
google-cloud-bigquery==3.14.1
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
Werkzeug==3.0.1
